moved the catch clause closer to the thrown exception
